Refactor modules to avoid import from scripts
